quantum-tools/docs/references/smartwater_research_agent/smartwater_research_agent/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
pydantic>=2.8.2
python-dotenv>=1.0.1
tenacity>=8.5.0
tqdm>=4.66.4
feedparser>=6.0.11
sqlite-utils>=3.37
networkx>=3.3
pandas>=2.2.2
